Rename ui state "lithology" to "well-log"
