+++ assets/src/scripts/components/graph/state/index.js
@@ -4,11 +4,11 @@ import options from './options';
 
 export { getCursor, getCursorDatum, setCursor } from './cursor';
 export * from './layout';
-export * from './lithology';
 export * from './options';
 export { getActiveClasses, getChartPoints, getCurrentWaterLevels,
          getCurrentWaterLevelUnit, getExtentX, getLineSegments } from './points';
 export * from './scales';
+export * from './well-log';
 
 export default {
     ...cursor,
