+++ assets/src/scripts/components/graph/view/construction.js
@@ -1,0 +1,6 @@
+export default function (elem, options, container) {
+    container = container || elem
+        .append('g');
+
+    return container;
+}
